Normed setcmd.c
